+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - `convert` function modified
 - Test system modified
+- Digits type changed to `list`
 ## [1.1] - 2025-11-10
 ### Added
 - English circled mode