+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- English keycap mode
 ### Changed
 - `convert` function modified
 - Test system modified
+- `README.md` updated
 - Digits type changed to `list`
 ## [1.1] - 2025-11-10
 ### Added