+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - English full stop mode
 - English comma mode
+- English segmented mode
 - Malayalam numeral system
 ## [1.2] - 2025-12-11
 ### Added